+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2019-09-30
+### Added
+- `authenticators` can now have options. If you make an authenticator an object you can pass netChainIds and options.
+- Scatter is now in the demo as one of the options
+
 ## [0.2.0] - 2019-09-30
 ### Changed
 - The net option is now an array named nets